% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohash_util.R
\name{decode_bbox}
\alias{decode_bbox}
\title{geohash 문자열을 위경도 사각형(bbox)으로 디코딩한다}
\usage{
decode_bbox(geohash)
}
\arguments{
\item{geohash}{geohash 문자열.}
}
\value{
c(lat_min, lat_max, lon_min, lon_max) 이름이 붙은 numeric 벡터.
}
\description{
geohash 문자열을 위경도 사각형(bbox)으로 디코딩한다
}
