% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohash_util.R
\name{decode_center}
\alias{decode_center}
\title{geohash 문자열을 중심 좌표로 디코딩한다}
\usage{
decode_center(geohash)
}
\arguments{
\item{geohash}{geohash 문자열.}
}
\value{
c(lat, lon) 이름이 붙은 numeric 벡터.
}
\description{
geohash 문자열을 중심 좌표로 디코딩한다
}
