% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run2.R
\name{run2}
\alias{run2}
\title{Rpuzzle 'JSON 합치기' GUI를 실행한다}
\usage{
run2()
}
\value{
(보이지 않게) \code{\link{puzzle_app2}}의 반환값.
}
\description{
유동인구/카드매출/상권 JSON 파일을 각각 2개씩 지정해 하나로 합칠 수 있는
팝업창을 띄운다. \code{\link{puzzle_app2}}를 호출한다.
}
