% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app2.R
\name{puzzle_app2}
\alias{puzzle_app2}
\title{Rpuzzle 'JSON 합치기' 팝업창을 띄운다}
\usage{
puzzle_app2()
}
\value{
(보이지 않게) 생성된 최상위 Tcl/Tk 창 핸들.
}
\description{
유동인구/카드매출/상권 JSON 파일을 각각 2개씩 지정하고 버튼을 클릭하면
두 파일의 데이터를 하나로 합친 "_merged.json" 파일을 생성하는
Tcl/Tk 팝업창을 띄운다.
}
